split shared-controller plus fix imports
