+++ api/v3/wms_types_test.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/google/go-cmp/cmp"
 	"github.com/pdok/smooth-operator/model"
-	controller "github.com/pdok/smooth-operator/pkg/util"
+	smoothoperatorutils "github.com/pdok/smooth-operator/pkg/util"
 )
 
 func TestLayer_setInheritedBoundingBoxes(t *testing.T) {
@@ -60,19 +60,19 @@ func TestLayer_setInheritedBoundingBoxes(t *testing.T) {
 		{
 			name: "setInheritedBoundingBoxes for layer",
 			layer: Layer{
-				Name:          controller.Pointer("toplayer"),
+				Name:          smoothoperatorutils.Pointer("toplayer"),
 				BoundingBoxes: []WMSBoundingBox{first28992BoundingBox},
 				Layers: []Layer{
 					{
-						Name:          controller.Pointer("grouplayer-1"),
+						Name:          smoothoperatorutils.Pointer("grouplayer-1"),
 						BoundingBoxes: []WMSBoundingBox{first4326BoundingBox},
 						Layers: []Layer{
 							{
-								Name:          controller.Pointer("datalayer-1"),
+								Name:          smoothoperatorutils.Pointer("datalayer-1"),
 								BoundingBoxes: []WMSBoundingBox{first4258BoundingBox},
 							},
 							{
-								Name:          controller.Pointer("datalayer-2"),
+								Name:          smoothoperatorutils.Pointer("datalayer-2"),
 								BoundingBoxes: []WMSBoundingBox{second28992BoundingBox},
 							},
 						},
@@ -129,9 +129,9 @@ func TestLayer_setInheritedBoundingBoxes(t *testing.T) {
 }
 
 func TestLayer_GetParent(t *testing.T) {
-	childLayer2 := Layer{Name: controller.Pointer("childlayer-2")}
-	childLayer1 := Layer{Name: controller.Pointer("childlayer-1"), Layers: []Layer{childLayer2}}
-	topLayer := Layer{Name: controller.Pointer("toplayer"), Layers: []Layer{childLayer1}}
+	childLayer2 := Layer{Name: smoothoperatorutils.Pointer("childlayer-2")}
+	childLayer1 := Layer{Name: smoothoperatorutils.Pointer("childlayer-1"), Layers: []Layer{childLayer2}}
+	topLayer := Layer{Name: smoothoperatorutils.Pointer("toplayer"), Layers: []Layer{childLayer1}}
 
 	type args struct {
 		service WMSService
