+++ cmd/main.go
@@ -29,7 +29,7 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/log/zap"
 
 	"github.com/pdok/mapserver-operator/internal/controller/mapfilegenerator"
-	smoothoperator "github.com/pdok/smooth-operator/api/v1"
+	smoothoperatorv1 "github.com/pdok/smooth-operator/api/v1"
 	traefikiov1alpha1 "github.com/traefik/traefik/v3/pkg/provider/kubernetes/crd/traefikio/v1alpha1"
 
 	// Import all Kubernetes client auth plugins (e.g. Azure, GCP, OIDC, etc.)
@@ -72,7 +72,7 @@ var (
 func init() {
 	utilruntime.Must(clientgoscheme.AddToScheme(scheme))
 	utilruntime.Must(traefikiov1alpha1.AddToScheme(scheme))
-	utilruntime.Must(smoothoperator.AddToScheme(scheme))
+	utilruntime.Must(smoothoperatorv1.AddToScheme(scheme))
 	utilruntime.Must(pdoknlv3.AddToScheme(scheme))
 	utilruntime.Must(pdoknlv2beta1.AddToScheme(scheme))
 	// +kubebuilder:scaffold:scheme
