+++ internal/controller/mapserver/deployment_test.go
@@ -8,7 +8,7 @@ import (
 	"github.com/pdok/mapserver-operator/internal/controller/types"
 	smoothoperatorutils "github.com/pdok/smooth-operator/pkg/util"
 	"github.com/stretchr/testify/assert"
-	v1 "k8s.io/api/core/v1"
+	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/api/resource"
 	"sigs.k8s.io/yaml"
 
@@ -23,7 +23,7 @@ func TestGetVolumeMountsForDeployment(t *testing.T) {
 	pdoknlv3.SetHost("https://service.pdok.nl")
 	result := GetVolumeMountsForDeployment(wfs, "/srv")
 
-	var expectedVolumeMounts struct{ VolumeMounts []v1.VolumeMount }
+	var expectedVolumeMounts struct{ VolumeMounts []corev1.VolumeMount }
 	err := yaml.Unmarshal(expectedVolumeMountsYaml, &expectedVolumeMounts)
 	assert.NoError(t, err)
 	assert.Equal(t, expectedVolumeMounts.VolumeMounts, result)
@@ -36,7 +36,7 @@ func TestGetEnvVarsForDeployment(t *testing.T) {
 	var wfs = getV3()
 	pdoknlv3.SetHost("https://service.pdok.nl")
 	result := GetEnvVarsForDeployment(wfs, "blobs-secret")
-	var expectedEnvVars struct{ EnvVars []v1.EnvVar }
+	var expectedEnvVars struct{ EnvVars []corev1.EnvVar }
 	err := yaml.Unmarshal(expectedEnvVarsYaml, &expectedEnvVars)
 	assert.NoError(t, err)
 	assert.Equal(t, expectedEnvVars.EnvVars, result)
@@ -47,16 +47,16 @@ func TestGetResourcesForDeployment(t *testing.T) {
 	pdoknlv3.SetHost("https://service.pdok.nl")
 	result := GetResourcesForDeployment(wfs)
 
-	expectedLimits := v1.ResourceList{}
-	expectedRequest := v1.ResourceList{}
+	expectedLimits := corev1.ResourceList{}
+	expectedRequest := corev1.ResourceList{}
 
-	expectedLimits[v1.ResourceMemory] = resource.MustParse("800M")
-	expectedLimits[v1.ResourceEphemeralStorage] = resource.MustParse("505Mi")
+	expectedLimits[corev1.ResourceMemory] = resource.MustParse("800M")
+	expectedLimits[corev1.ResourceEphemeralStorage] = resource.MustParse("505Mi")
 
-	expectedRequest[v1.ResourceCPU] = resource.MustParse("0.15")
-	expectedRequest[v1.ResourceEphemeralStorage] = resource.MustParse("255Mi")
+	expectedRequest[corev1.ResourceCPU] = resource.MustParse("0.15")
+	expectedRequest[corev1.ResourceEphemeralStorage] = resource.MustParse("255Mi")
 
-	var expected = v1.ResourceRequirements{
+	var expected = corev1.ResourceRequirements{
 		Limits:   expectedLimits,
 		Requests: expectedRequest,
 		Claims:   nil,
@@ -80,9 +80,9 @@ func TestGetProbesForDeployment(t *testing.T) {
 	livenessResult, readinessResult, startupResult, err := GetProbesForDeployment(wfs)
 	assert.NoError(t, err)
 
-	var expectedLiveness v1.Probe
-	var expectedReadiness v1.Probe
-	var expectedStartup v1.Probe
+	var expectedLiveness corev1.Probe
+	var expectedReadiness corev1.Probe
+	var expectedStartup corev1.Probe
 	err = yaml.Unmarshal(expectedLivenessProbe, &expectedLiveness)
 	assert.NoError(t, err)
 	err = yaml.Unmarshal(expectedReadinessProbe, &expectedReadiness)
@@ -110,12 +110,12 @@ func TestGetVolumesForDeployment(t *testing.T) {
 	}
 	result := GetVolumesForDeployment(wfs, hashedConfigMapNames)
 
-	expected := []v1.Volume{
-		{Name: "base", VolumeSource: v1.VolumeSource{EmptyDir: &v1.EmptyDirVolumeSource{}}},
-		{Name: "data", VolumeSource: v1.VolumeSource{EmptyDir: &v1.EmptyDirVolumeSource{}}},
-		{Name: "mapserver", VolumeSource: v1.VolumeSource{ConfigMap: &v1.ConfigMapVolumeSource{LocalObjectReference: v1.LocalObjectReference{Name: "rws-nwbwegen-v1-0-wfs-mapserver-bb59c7f4f4"}, DefaultMode: smoothoperatorutils.Pointer(int32(420))}}},
-		{Name: "capabilities-generator-config", VolumeSource: v1.VolumeSource{ConfigMap: &v1.ConfigMapVolumeSource{LocalObjectReference: v1.LocalObjectReference{Name: "rws-nwbwegen-v1-0-wfs-capabilities-generator-6m4mfkgb5d"}, DefaultMode: smoothoperatorutils.Pointer(int32(420))}}},
-		{Name: "mapfile-generator-config", VolumeSource: v1.VolumeSource{ConfigMap: &v1.ConfigMapVolumeSource{LocalObjectReference: v1.LocalObjectReference{Name: "rws-nwbwegen-v1-0-wfs-mapfile-generator-bbbtd999dh"}, DefaultMode: smoothoperatorutils.Pointer(int32(420))}}},
+	expected := []corev1.Volume{
+		{Name: "base", VolumeSource: corev1.VolumeSource{EmptyDir: &corev1.EmptyDirVolumeSource{}}},
+		{Name: "data", VolumeSource: corev1.VolumeSource{EmptyDir: &corev1.EmptyDirVolumeSource{}}},
+		{Name: "mapserver", VolumeSource: corev1.VolumeSource{ConfigMap: &corev1.ConfigMapVolumeSource{LocalObjectReference: corev1.LocalObjectReference{Name: "rws-nwbwegen-v1-0-wfs-mapserver-bb59c7f4f4"}, DefaultMode: smoothoperatorutils.Pointer(int32(420))}}},
+		{Name: "capabilities-generator-config", VolumeSource: corev1.VolumeSource{ConfigMap: &corev1.ConfigMapVolumeSource{LocalObjectReference: corev1.LocalObjectReference{Name: "rws-nwbwegen-v1-0-wfs-capabilities-generator-6m4mfkgb5d"}, DefaultMode: smoothoperatorutils.Pointer(int32(420))}}},
+		{Name: "mapfile-generator-config", VolumeSource: corev1.VolumeSource{ConfigMap: &corev1.ConfigMapVolumeSource{LocalObjectReference: corev1.LocalObjectReference{Name: "rws-nwbwegen-v1-0-wfs-mapfile-generator-bbbtd999dh"}, DefaultMode: smoothoperatorutils.Pointer(int32(420))}}},
 	}
 
 	assert.Equal(t, expected, result)
