+++ internal/controller/middleware.go
@@ -1,0 +1,48 @@
+package controller
+
+import (
+	pdoknlv3 "github.com/pdok/mapserver-operator/api/v3"
+	smoothoperatorutils "github.com/pdok/smooth-operator/pkg/util"
+	"github.com/traefik/traefik/v3/pkg/config/dynamic"
+	traefikiov1alpha1 "github.com/traefik/traefik/v3/pkg/provider/kubernetes/crd/traefikio/v1alpha1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	ctrl "sigs.k8s.io/controller-runtime"
+)
+
+const corsHeadersName = "mapserver-headers"
+
+func getBareCorsHeadersMiddleware[O pdoknlv3.WMSWFS](obj O) *traefikiov1alpha1.Middleware {
+	return &traefikiov1alpha1.Middleware{
+		ObjectMeta: metav1.ObjectMeta{
+			Name: getSuffixedName(obj, corsHeadersName),
+			// name might become too long. not handling here. will just fail on apply.
+			Namespace: obj.GetNamespace(),
+			UID:       obj.GetUID(),
+		},
+	}
+}
+
+func mutateCorsHeadersMiddleware[R Reconciler, O pdoknlv3.WMSWFS](r R, obj O, middleware *traefikiov1alpha1.Middleware) error {
+	reconcilerClient := getReconcilerClient(r)
+
+	labels := addCommonLabels(obj, smoothoperatorutils.CloneOrEmptyMap(obj.GetLabels()))
+	if err := smoothoperatorutils.SetImmutableLabels(reconcilerClient, middleware, labels); err != nil {
+		return err
+	}
+	middleware.Spec = traefikiov1alpha1.MiddlewareSpec{
+		Headers: &dynamic.Headers{
+			CustomResponseHeaders: map[string]string{
+				"Access-Control-Allow-Headers": "Content-Type",
+				"Access-Control-Allow-Method":  "GET, POST, OPTIONS",
+				"Access-Control-Allow-Origin":  "*",
+				"Cache-Control":                "public, max-age=3600, no-transform",
+			},
+		},
+	}
+
+	if err := smoothoperatorutils.EnsureSetGVK(reconcilerClient, middleware, middleware); err != nil {
+		return err
+	}
+
+	return ctrl.SetControllerReference(obj, middleware, getReconcilerScheme(r))
+}
