+++ internal/controller/poddisruptionbudget.go
@@ -1,0 +1,41 @@
+package controller
+
+import (
+	pdoknlv3 "github.com/pdok/mapserver-operator/api/v3"
+	smoothoperatorutils "github.com/pdok/smooth-operator/pkg/util"
+	policyv1 "k8s.io/api/policy/v1"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/util/intstr"
+	ctrl "sigs.k8s.io/controller-runtime"
+)
+
+func getBarePodDisruptionBudget[O pdoknlv3.WMSWFS](obj O) *policyv1.PodDisruptionBudget {
+	return &policyv1.PodDisruptionBudget{
+		ObjectMeta: metav1.ObjectMeta{
+			Name:      getSuffixedName(obj, MapserverName),
+			Namespace: obj.GetNamespace(),
+		},
+	}
+}
+
+func mutatePodDisruptionBudget[R Reconciler, O pdoknlv3.WMSWFS](r R, obj O, podDisruptionBudget *policyv1.PodDisruptionBudget) error {
+	reconcilerClient := getReconcilerClient(r)
+
+	labels := addCommonLabels(obj, smoothoperatorutils.CloneOrEmptyMap(obj.GetLabels()))
+	if err := smoothoperatorutils.SetImmutableLabels(reconcilerClient, podDisruptionBudget, labels); err != nil {
+		return err
+	}
+
+	matchLabels := smoothoperatorutils.CloneOrEmptyMap(labels)
+	podDisruptionBudget.Spec = policyv1.PodDisruptionBudgetSpec{
+		MaxUnavailable: &intstr.IntOrString{Type: intstr.Int, IntVal: 1},
+		Selector: &metav1.LabelSelector{
+			MatchLabels: matchLabels,
+		},
+	}
+
+	if err := smoothoperatorutils.EnsureSetGVK(reconcilerClient, podDisruptionBudget, podDisruptionBudget); err != nil {
+		return err
+	}
+	return ctrl.SetControllerReference(obj, podDisruptionBudget, getReconcilerScheme(r))
+}
